Break out OtlpIntegrationTest helper
