+++ contrib/jmx-metrics/src/test/groovy/io/opentelemetry/contrib/jmxmetrics/OtlpIntegrationTests.groovy
@@ -16,13 +16,6 @@
 
 package io.opentelemetry.contrib.jmxmetrics
 
-import static org.junit.Assert.assertTrue
-
-import io.grpc.ServerBuilder
-import io.grpc.stub.StreamObserver
-import io.opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceRequest
-import io.opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceResponse
-import io.opentelemetry.proto.collector.metrics.v1.MetricsServiceGrpc
 import io.opentelemetry.proto.common.v1.InstrumentationLibrary
 import io.opentelemetry.proto.common.v1.StringKeyValue
 import io.opentelemetry.proto.metrics.v1.InstrumentationLibraryMetrics
@@ -30,10 +23,8 @@ import io.opentelemetry.proto.metrics.v1.Metric
 import io.opentelemetry.proto.metrics.v1.DoubleHistogram
 import io.opentelemetry.proto.metrics.v1.DoubleHistogramDataPoint
 import io.opentelemetry.proto.metrics.v1.ResourceMetrics
-import java.util.concurrent.TimeUnit
 import org.testcontainers.Testcontainers
 import spock.lang.Requires
-import spock.lang.Shared
 import spock.lang.Timeout
 import spock.lang.Unroll
 
@@ -41,34 +32,7 @@ import spock.lang.Unroll
     System.getProperty('ojc.integration.tests') == 'true'
 })
 @Timeout(60)
-class OtlpIntegrationTests extends IntegrationTest  {
-
-    @Shared
-    def collector
-    @Shared
-    def collectorServer
-    @Shared
-    def otlpPort
-
-    def setup() {
-        // set up a collector per test to avoid noisy neighbor
-        otlpPort = availablePort()
-        collector = new Collector()
-        collectorServer = ServerBuilder.forPort(otlpPort).addService(collector).build()
-        collectorServer.start()
-    }
-
-    def cleanup() {
-        collectorServer.shutdownNow()
-        collectorServer.awaitTermination(5, TimeUnit.SECONDS)
-    }
-
-    def availablePort() {
-        def sock = new ServerSocket(0);
-        def port = sock.getLocalPort()
-        sock.close()
-        return port
-    }
+class OtlpIntegrationTests extends OtlpIntegrationTest {
 
     @Unroll
     def 'end to end with stdin config: #useStdin'() {
@@ -134,41 +98,5 @@ class OtlpIntegrationTests extends IntegrationTest  {
         useStdin | _
         false | _
         true | _
-    }
-
-    static final class Collector extends MetricsServiceGrpc.MetricsServiceImplBase {
-        private final List<ResourceMetrics> receivedMetrics = new ArrayList<>()
-        private final Object monitor = new Object()
-
-        @Override
-        void export(
-                ExportMetricsServiceRequest request,
-                StreamObserver<ExportMetricsServiceResponse> responseObserver) {
-            synchronized (receivedMetrics) {
-                receivedMetrics.addAll(request.resourceMetricsList)
-            }
-            synchronized (monitor) {
-                monitor.notify()
-            }
-            responseObserver.onNext(ExportMetricsServiceResponse.newBuilder().build())
-            responseObserver.onCompleted()
-        }
-
-        List<ResourceMetrics> getReceivedMetrics() {
-            List<ResourceMetrics> received
-            try {
-                synchronized (monitor) {
-                    monitor.wait(15000)
-                }
-            } catch (final InterruptedException e) {
-                assertTrue(e.message, false)
-            }
-
-            synchronized (receivedMetrics) {
-                received = new ArrayList<>(receivedMetrics)
-                receivedMetrics.clear()
-            }
-            return received
-        }
     }
 }
